new entities added

+++ src/main/java/com/bharath/flightreservation/repos/AppointmentDataRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface AppointmentDataRepository extends JpaRepository<Appointment_data, Long> {
+public interface AppointmentDataRepository extends JpaRepository<Appointment_data, Integer> {
 
 
 }
